requirements.txt
ctransformers==0..2.5
sentence-transformers==2.2.2
pinecone-client
langchain==0.0.225
langchain_pinecone
flask
pypdf
python-dotenv
-e .